% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_as_is.R
\name{fac_as_is}
\alias{fac_as_is}
\title{Factors as is}
\usage{
fac_as_is(x, ...)
}
\arguments{
\item{x}{a vector of data}

\item{...}{Name-value pairs of summary functions like factor()}
}
\value{
factor
}
\description{
Wrapped around factor, factor as is assignes the levels of the factor in the order in which they appear in the data.
}
\examples{
a <- c('sam', 'harry', 'andy')
fac_as_is(a)
factor(a)
}

